Find well fitting admixture graphs

This function generates and evaluates admixture graphs in numgen iterations to find well fitting admixturegraphs.

find_graphs2(
  f2_blocks,
  numgen = 1,
  numgraphs = 10,
  numadmix = 0,
  stopscore = 0,
  stop_after = NULL,
  stopafter_noimprovement = 15,
  initgraph = NULL,
  outpop = NULL,
  mutfuns = namedList(spr_leaves, spr_all, swap_leaves, move_admixedge_once,
    flipadmix_random, mutate_n),
  opt_worst_residual = FALSE,
  plusminus_generations = 5,
  admix_constraints = NULL,
  event_constraints = NULL,
  reject_f4z = 0,
  verbose = TRUE,
  ...
)

Arguments

f2_blocks
A 3d array of blocked f2 statistics, output of f2_from_precomp or extract_f2

numgen
Number of generations

numgraphs
Number of graphs in each generation

numadmix
Number of admixture events within each graph. (Only relevant if initgraph = NULL)

mutfuns
Functions used to modify graphs. Defaults to the following:

spr_leaves: Subtree prune and regraft leaves. Cuts a leaf node and attaches it to a random other edge in the graph.

spr_all: Subtree prune and regraft. Cuts any edge and attaches the new orphan node to a random other edge in the graph, keeping the number of admixture nodes constant.

swap_leaves: Swaps two leaf nodes.

move_admixedge_once: Moves an admixture edge to a nearby location.

flipadmix_random: Flips the direction of an admixture edge (if possible).

mutate_n: Apply n of the mutation functions in this list to a graph (defaults to 2).

opt_worst_residual
Optimize for lowest worst residual instead of best score. FALSE by default, because the likelihood score is generally a better indicator of the quality of the model fit, and because optimizing for the lowest worst residual is slower (because f4-statistics need to be computed).

plusminus_generations
If the best score does not improve after plusminus_generations generations, another approach to improving the score will be attempted: A number of graphs with on additional admixture edge will be generated and evaluated. The resulting graph with the best score will be picked, and new graphs will be created by removing any one admixture edge (bringing the number back to what it was originally). The graph with the lowest score will then be selected. This often makes it possible to break out of local optima, but is slower than regular graph modifications.

admix_constraints
A data frame with constraints on the number of admixture events for each population. See satisfies_numadmix As soon as one graph happens to satisfy these constraints, all subsequently generated graphs will be required to also satisfy them.

event_constraints
A data frame with constraints on the order of events in an admixture graph. See satisfies_eventorder As soon as one graph happens to satisfy these constraints, all subsequently generated graphs will be required to also satisfy them.

reject_f4z
If this is a number greater than zero, all f4-statistics with z > reject_f4z will be used to constrain the search space of admixture graphs: Any graphs in which any of the relevant f4-statistics are expected to be zero will not be evaluated.

verbose
Print progress updates

...
Additional arguments passed to qpgraph

Value

A nested data frame with one model per line

See also

Examples

if (FALSE) {
pops = dimnames(example_f2_blocks)[[1]]
initgraph = random_admixturegraph(pops, 2, outpop = 'Chimp.REF')
res = find_graphs2(example_f2_blocks, initgraph, numgen = 100)
res %>% slice_min(score)
}
if (FALSE) {
# Making new mutation functions by modifying or combining existing ones:
newfun1 = function(graph, ...) mutate_n(graph, 3, ...)
newfun2 = function(graph, ...) flipadmix_random(spr_leaves(graph, ...), ...)
find_graphs(f2_blocks, mutfuns = namedList(spr_leaves, newfun1, newfun2), mutprobs = c(0.2, 0.3, 0.5))
}