Estimate admixture weights

Models target as a mixture of left populations, and outgroup right populations. Uses Lazaridis method based non-negative least squares of f4 matrix.

lazadm(data, left, right, target, boot = FALSE, constrained = TRUE)

Arguments

data
The input data in the form of:

A 3d array of blocked f2 statistics, output of f2_from_precomp or extract_f2

A directory with f2 statistics

The prefix of a genotype file

left
Left populations (sources)

right
Right populations (outgroups)

target
Target population

boot
If FALSE (the default), each block will be left out at a time and the covariance matrix of f4 statistics, as well as the weight standard errors, will be computed using block-jackknife. Otherwise bootstrap resampling is performed n times, where n is either equal to boot if it is an integer, or equal to the number of blocks if boot is TRUE. The the covariance matrix of f4 statistics, as well as the weight standard errors, will be computed using bootstrapping.

constrained
Constrain admixture weights to be non-negative

Value

lazadm returns a data frame with weights and standard errors for each left population

References

Patterson, N. et al. (2012) Ancient admixture in human history. Genetics

Haak, W. et al. (2015) Massive migration from the steppe was a source for Indo-European languages in Europe. Nature (SI 9)

See also

Examples

target = 'Denisova.DG'
left = c('Altai_Neanderthal.DG', 'Vindija.DG')
right = c('Chimp.REF', 'Mbuti.DG', 'Russia_Ust_Ishim.DG', 'Switzerland_Bichon.SG')
lazadm(example_f2_blocks, left, right, target)

#> # A tibble: 2 x 5
#>   target      left                   weight       se       z
#>   <chr>       <chr>                   <dbl>    <dbl>   <dbl>
#> 1 Denisova.DG Altai_Neanderthal.DG 1.00e+ 0 1.94e-13 5.17e12
#> 2 Denisova.DG Vindija.DG           2.49e-12 1.93e-13 1.29e 1

lazadm(example_f2_blocks, left, right, target, constrained = FALSE)

#> # A tibble: 2 x 5
#>   target      left                 weight    se      z
#>   <chr>       <chr>                 <dbl> <dbl>  <dbl>
#> 1 Denisova.DG Altai_Neanderthal.DG   4.79  6.69  0.716
#> 2 Denisova.DG Vindija.DG            -3.79  6.69 -0.567