Plot an admixture graph on a map

Plot an admixture graph on a map

plot_graph_map(graph, leafcoords, shapedata = NULL)

Arguments

graph

a two column matrix specifying the admixture graph. first column is source node, second column is target node. the first edge has to be root -> outgroup. admixture nodes are inferred as those nodes which are the targets of two different sources.

leafcoords

data frame with columns group, lon, lat

shapedata

shapedata

Value

a plotly object

Examples

plot_graph_map(example_igraph, example_anno)