Plot an admixture graph using plotly

Plot an admixture graph using plotly

plotly_graph(
  graph,
  collapse_threshold = 0,
  fix = FALSE,
  print_highlow = FALSE,
  highlight_unidentifiable = FALSE,
  pos = NULL,
  nudge_y = -0.1,
  annot = ""
)

Arguments

graph

An admixture graph

collapse_threshold

Collapse nodes if they are separated by less than this distance (for fitted graphs)

fix

If TRUE, there will be an attempt to rearrange the nodes to minimize the number of intersecting edges. This can take very long for large graphs. By default this is only done for graphs with fewer than 10 leaves.

highlight_unidentifiable

Highlight unidentifiable edges in red. Can be slow for large graphs. See unidentifiable_edges.

pos

Optional data frame with node coordinates (columns node, x, y)

Value

A plotly object

Examples

plotly_graph(example_graph)