Faster version of qpadm with reduced output

Models target as a mixture of left populations, given a set of outgroup right populations. Can be used to estimate admixture proportions, and to estimate the number of independent admixture events.

qpadm_p(
  f2_data,
  left,
  right,
  target = NULL,
  fudge = 1e-04,
  boot = FALSE,
  constrained = FALSE,
  rnk = length(setdiff(left, target)) - 1,
  cpp = TRUE
)

Arguments

left
Left populations (sources)

right
Right populations (outgroups)

target
Target population

fudge
value added to diagonal matrix elements before inverting

boot
If FALSE (the default), each block will be left out at a time and the covariance matrix of f4 statistics, as well as the weight standard errors, will be computed using block-jackknife. Otherwise bootstrap resampling is performed n times, where n is either equal to boot if it is an integer, or equal to the number of blocks if boot is TRUE. The the covariance matrix of f4 statistics, as well as the weight standard errors, will be computed using bootstrapping.

constrained
Constrain admixture weights to be non-negative

rnk
Rank of f4-matrix. Defaults to one less than full rank.

cpp
Use C++ functions. Setting this to FALSE will be slower but can help with debugging.

Value

Data frame with f4rank, dof, chisq, p, feasible

See also

Examples

left = c('Altai_Neanderthal.DG', 'Vindija.DG')
right = c('Chimp.REF', 'Mbuti.DG', 'Russia_Ust_Ishim.DG', 'Switzerland_Bichon.SG')
target = 'Denisova.DG'
qpadm_p(example_f2_blocks, left, right, target)

#> # A tibble: 1 x 5
#>   f4rank   dof chisq      p feasible
#>    <dbl> <dbl> <dbl>  <dbl> <lgl>   
#> 1      1     2  7.15 0.0280 FALSE