Compute f3-statistics from f2-statistics.

Takes a 3d array of f2 block jackknife estimates and computes f3-statistics between the first population p1 and all population pairs i, j: f3(p1; p_i, p_j)

qpgraph_precompute_f3(pops, f2_blocks = NULL, block_lengths = NULL,
  f2_dir = NULL, f2_denom = 1, fudge = 0.001, lsqmode = FALSE)

Arguments

pops
populations for which to compute f3-statistics

f2_blocks
3d array of block-jackknife leave-one-block-out estimates of f2 statistics. output of afs_to_f2_blocks. they are weighted by inverse of outgroup heterozygosity, if outgroup was specified.

block_lengths
the jackknife block lengths used in computing the f2 statistics. see get_block_lengths.

f2_dir
a directory with f2 statistics for each population pair in the graph. must contain 'block_lengths.RData'.

f2_denom
scales f2-statistics. A value of around 0.278 converts F2 to Fst.

fudge
try increasing this, if you get the error message constraints are inconsistent, no solution!.

lsqmode
least-squares mode. sets the offdiagonal elements of the block-jackknife covariance matrix to zero.

Value

a list with four items

f3_jest a matrix with f3-statistics for all population pairs with the output

ppinv a matrix with the inverse of the f3-statistic covariance matrix

f2out a data frame with f2 estimates

f3out a data frame with f3 estimates

Examples

pops = get_leafnames(example_igraph)
qpgraph_precompute_f3(pops, example_f2_blocks, example_block_lengths)

#> $f3_jest
#>  [1] 0.17485985 0.17116794 0.09600898 0.11127524 0.09724603 0.09717553
#>  [7] 0.17760530 0.09610525 0.11103233 0.09759183 0.09750339 0.15957981
#> [13] 0.08594492 0.15026916 0.15129538 0.15717535 0.08543411 0.08553202
#> [19] 0.19965759 0.19638099 0.34035924
#> 
#> $ppinv
#>               [,1]         [,2]         [,3]          [,4]         [,5]
#>  [1,] 15341009.160  -8629845.15  -2824645.79   -251300.599  -3242426.80
#>  [2,] -8629845.150  23148218.32   -380971.73   -479336.672   -633543.99
#>  [3,] -2824645.788   -380971.73  43619562.92  -1519402.984  -8717267.55
#>  [4,]  -251300.599   -479336.67  -1519402.98  29973440.648  -1283463.23
#>  [5,] -3242426.804   -633543.99  -8717267.55  -1283463.227  36925741.86
#>  [6,] -3064410.116   -391211.94  -7374492.78   -890001.331  -8218182.77
#>  [7,] -4814930.357 -10816218.12   2044064.89    334438.749   2900410.93
#>  [8,]  2001123.199   -567798.28 -30513398.38   1117501.346   5261007.64
#>  [9,]   -55764.460  -1026689.37   1426999.32 -26623786.510   2220572.22
#> [10,]  1722094.342    375707.96   4891195.98    734746.611 -28472176.02
#> [11,]  2478500.719   -721617.03   5005283.04    496423.014   5453938.69
#> [12,]     3630.619    178472.34   -436796.19    323105.946    388815.39
#> [13,]   998813.356    274639.16  -6648193.47  -1443193.293   1145920.49
#> [14,]   -89534.968   -110310.41  -1290295.22    300041.910     21219.18
#> [15,]    34499.202    304855.75    365678.45      6332.733    109088.81
#> [16,]  -527248.909   -911337.88   -259387.33   -210033.953    -45949.30
#> [17,]   -31851.564    313180.63   2975621.59   -628154.358  -4815874.92
#> [18,]   108236.766     75690.93    520173.69   -670138.369   1690039.17
#> [19,]  -128093.924     67728.40    -26755.02      -256.245   -113610.15
#> [20,]   -23218.583    -95727.60    -37265.81    -55979.006    412943.18
#> [21,]  -197408.455    348904.88    -84680.75    399738.352   -126469.56
#>               [,6]         [,7]          [,8]         [,9]       [,10]
#>  [1,]  -3064410.12  -4814930.36  2.001123e+06    -55764.46   1722094.3
#>  [2,]   -391211.94 -10816218.12 -5.677983e+05  -1026689.37    375708.0
#>  [3,]  -7374492.78   2044064.89 -3.051340e+07   1426999.32   4891196.0
#>  [4,]   -890001.33    334438.75  1.117501e+06 -26623786.51    734746.6
#>  [5,]  -8218182.77   2900410.93  5.261008e+06   2220572.22 -28472176.0
#>  [6,]  36108598.33   1522163.97  3.867309e+06   1467804.10   5531691.3
#>  [7,]   1522163.97  16779739.33 -2.488812e+06   -350329.79  -2958293.8
#>  [8,]   3867308.54  -2488811.91  4.201596e+07  -1636719.10  -7751926.0
#>  [9,]   1467804.10   -350329.79 -1.636719e+06  31017622.15  -1020712.6
#> [10,]   5531691.31  -2958293.78 -7.751926e+06  -1020712.60  35757975.1
#> [11,] -26306495.72  -3088565.04 -7.364749e+06   -695173.31  -7688246.2
#> [12,]     43807.35   -561801.83  6.254054e+04   -820011.80  -1058756.4
#> [13,]   1508751.80    830706.22 -5.913982e+06   -864101.50   1249202.3
#> [14,]   -527887.18    -31279.69 -7.513241e+05     74712.66   2255862.8
#> [15,]   -524918.84    425889.39 -3.711365e+05    -21126.27  -1363774.6
#> [16,]    531095.32  -1929581.81  1.240552e+06  -1239846.49   1057974.7
#> [17,]   1397893.67   1076031.63  1.209656e+06   -993277.66  -5993111.2
#> [18,]  -4538708.91    478349.13  1.718205e+06     70689.36   2381423.5
#> [19,]   -185448.21     46317.55  1.069228e+04   -139590.58    354820.8
#> [20,]    291958.79    284240.03 -3.216003e+05     97240.03    145903.0
#> [21,]    -90888.03    -40371.51  7.069008e+02   -152883.44   -380970.8
#>              [,11]        [,12]       [,13]       [,14]        [,15]
#>  [1,]   2478500.72     3630.619   998813.36   -89534.97    34499.202
#>  [2,]   -721617.03   178472.339   274639.16  -110310.41   304855.748
#>  [3,]   5005283.04  -436796.193 -6648193.47 -1290295.22   365678.452
#>  [4,]    496423.01   323105.946 -1443193.29   300041.91     6332.733
#>  [5,]   5453938.69   388815.390  1145920.49    21219.18   109088.807
#>  [6,] -26306495.72    43807.347  1508751.80  -527887.18  -524918.845
#>  [7,]  -3088565.04  -561801.831   830706.22   -31279.69   425889.389
#>  [8,]  -7364749.01    62540.543 -5913982.32  -751324.10  -371136.495
#>  [9,]   -695173.31  -820011.799  -864101.50    74712.66   -21126.272
#> [10,]  -7688246.24 -1058756.450  1249202.29  2255862.83 -1363774.565
#> [11,]  33498610.77  -364724.441  1040370.88  -516107.16    76930.097
#> [12,]   -364724.44  9892773.965 -1585031.10 -3436287.69 -2862293.612
#> [13,]   1040370.88 -1585031.102 29402662.12   276982.59   868793.633
#> [14,]   -516107.16 -3436287.688   276982.59 11830239.69 -5206034.740
#> [15,]     76930.10 -2862293.612   868793.63 -5206034.74 12812190.803
#> [16,]   1125474.02    55689.640 -4898289.95  -179725.51    17123.709
#> [17,]   1243097.76   803976.117 -7272798.34 -2273467.14   915786.548
#> [18,]  -3667933.00   616164.647 -6072004.00  1221637.41 -1149801.418
#> [19,]     45065.45   -15612.148    71825.15  -313792.77    19937.742
#> [20,]   -241547.35   310815.817  -142165.26 -1388116.20    85981.253
#> [21,]    250093.07   -19242.439   463353.53   837515.09 -3439441.745
#>             [,16]       [,17]       [,18]       [,19]       [,20]         [,21]
#>  [1,]  -527248.91   -31851.56   108236.77 -128093.924   -23218.58  -197408.4551
#>  [2,]  -911337.88   313180.63    75690.93   67728.402   -95727.60   348904.8793
#>  [3,]  -259387.33  2975621.59   520173.69  -26755.020   -37265.81   -84680.7541
#>  [4,]  -210033.95  -628154.36  -670138.37    -256.245   -55979.01   399738.3519
#>  [5,]   -45949.30 -4815874.92  1690039.17 -113610.147   412943.18  -126469.5575
#>  [6,]   531095.32  1397893.67 -4538708.91 -185448.213   291958.79   -90888.0254
#>  [7,] -1929581.81  1076031.63   478349.13   46317.549   284240.03   -40371.5117
#>  [8,]  1240552.50  1209655.61  1718205.29   10692.282  -321600.29      706.9008
#>  [9,] -1239846.49  -993277.66    70689.36 -139590.582    97240.03  -152883.4407
#> [10,]  1057974.69 -5993111.24  2381423.54  354820.787   145903.04  -380970.7983
#> [11,]  1125474.02  1243097.76 -3667933.00   45065.454  -241547.35   250093.0720
#> [12,]    55689.64   803976.12   616164.65  -15612.148   310815.82   -19242.4390
#> [13,] -4898289.95 -7272798.34 -6072004.00   71825.153  -142165.26   463353.5309
#> [14,]  -179725.51 -2273467.14  1221637.41 -313792.774 -1388116.20   837515.0928
#> [15,]    17123.71   915786.55 -1149801.42   19937.742    85981.25 -3439441.7449
#> [16,]  9305421.79 -2819367.85 -2703670.80  -13786.415    11443.07  -421541.1377
#> [17,] -2819367.85 24282777.17 -7664940.23   73557.157  -314019.95   438598.8715
#> [18,] -2703670.80 -7664940.23 19812359.38   96786.278   269398.28 -1100617.9688
#> [19,]   -13786.42    73557.16    96786.28  594998.284  -502984.76   196496.3178
#> [20,]    11443.07  -314019.95   269398.28 -502984.757  2978958.32 -1677098.6507
#> [21,]  -421541.14   438598.87 -1100617.97  196496.318 -1677098.65  4523828.7218
#> 
#> $f2out
#> # A tibble: 21 x 4
#>    pop1                 pop2                   f2est       se
#>    <chr>                <chr>                  <dbl>    <dbl>
#>  1 Chimp.REF            Altai_Neanderthal.DG  0.175  0.000861
#>  2 Chimp.REF            Vindija.DG            0.178  0.000881
#>  3 Chimp.REF            Mbuti.DG              0.160  0.000477
#>  4 Chimp.REF            Denisova.DG           0.157  0.000771
#>  5 Chimp.REF            Russia_Ust_Ishim.DG   0.200  0.00152 
#>  6 Chimp.REF            Switzerland_Bichon.SG 0.340  0.000744
#>  7 Altai_Neanderthal.DG Vindija.DG            0.0101 0.000367
#>  8 Altai_Neanderthal.DG Mbuti.DG              0.142  0.000694
#>  9 Altai_Neanderthal.DG Denisova.DG           0.109  0.000933
#> 10 Altai_Neanderthal.DG Russia_Ust_Ishim.DG   0.180  0.00176 
#> # … with 11 more rows
#> 
#> $f3out
#> # A tibble: 21 x 4
#>    pop1                 pop2                  f3est       se
#>    <chr>                <chr>                 <dbl>    <dbl>
#>  1 Chimp.REF            Chimp.REF            0.175  0.000861
#>  2 Chimp.REF            Altai_Neanderthal.DG 0.171  0.000824
#>  3 Chimp.REF            Vindija.DG           0.0960 0.000473
#>  4 Chimp.REF            Mbuti.DG             0.111  0.000650
#>  5 Chimp.REF            Denisova.DG          0.0972 0.000550
#>  6 Chimp.REF            Russia_Ust_Ishim.DG  0.0972 0.000539
#>  7 Altai_Neanderthal.DG Altai_Neanderthal.DG 0.178  0.000881
#>  8 Altai_Neanderthal.DG Vindija.DG           0.0961 0.000471
#>  9 Altai_Neanderthal.DG Mbuti.DG             0.111  0.000643
#> 10 Altai_Neanderthal.DG Denisova.DG          0.0976 0.000557
#> # … with 11 more rows
#> 

if (FALSE) {
qpgraph_precompute_f3(pops, f2_dir = f2_dir, f2_denom = 0.278)
}