Compute the fit of an admixturegraph.

Computes the fit of an admixture graph for a given graph topology given f3 estimates and an inverted covariance matrix.

qpgraph_slim(graph, f3_jest, ppinv, fnscale = 1e-06, numstart = 10,
  seed = NULL, verbose = FALSE, cpp = TRUE)

Arguments

graph
an admixture graph represented as an igraph object

f3_jest
matrix of f3 estimates. output of qpgraph_precompute_f3

ppinv
inverted covariance matrix of f3-statistics

fnscale
try increasing or decreasing this, if optimization does not converge.

numstart
number of random initializations. defaults to 10 times the number of admixture nodes.

seed
seed for generating starting weights.

verbose
print optimization iterations

cpp
should optimization be run in C++ or in R? C++ is faster.

fudge
try increasing this, if you get the error message constraints are inconsistent, no solution!.

Value

a list of qpGraph output data

See also

Examples

pops = get_leafnames(example_igraph)
precomp = qpgraph_precompute_f3(pops, example_f2_blocks)

#> Error in get_f2(f2_data, pops, f2_denom): !is.character(f2_data) || dir.exists(f2_data) is not TRUE

f3_jest = precomp$f3_jest

#> Error in eval(expr, envir, enclos): object 'precomp' not found

ppinv = precomp$ppinv

#> Error in eval(expr, envir, enclos): object 'precomp' not found

out = qpgraph_slim(example_igraph, f3_jest, ppinv)

#> Error in qpgraph_slim(example_igraph, f3_jest, ppinv): object 'f3_jest' not found

plot_graph(out$edges)

#> Error in plot_graph(out$edges): object 'out' not found