Compute all pairwise f2 statistics

This function takes a allele frequency data and computes blocked f2 statistics for all population pairs, which are written to outdir. \(f2\) for each SNP is computed as \((p1 - p2)^2 - p1 (1 - p1)/(n1 - 1) - p2 (1 - p2)/(n2 - 1)\), where \(p1\) and \(p2\) are allele frequencies in populations \(1\) and \(2\), and \(n1\) and \(n2\) is the average number of non-missing haplotypes in populations \(1\) and \(2\). See details

afs_to_f2_blocks(
  afdat,
  maxmem = 8000,
  dist = 0.05,
  poly_only = TRUE,
  outdir = NULL,
  overwrite = FALSE,
  verbose = TRUE
)

Arguments

afdat
A list with three items with the same SNP in each row (generated by packedancestrymap_to_aftable or plink_to_aftable)

afs A matrix of allele frequencies for all populations (columns) and SNPs (rows)

counts A matrix of allele counts for all populations (columns) and SNPs (rows)

snpdat A data frame with SNP metadata

has to match nrow(afmat). See get_block_lengths

maxmem
split up allele frequency data into blocks, if memory requirements exceed maxmem MB.

dist
Genetic distance in Morgan. Default is 0.05 (50 cM).

poly_only
Exclude sites with identical allele frequencies in all populations

outdir
Directory into which to write f2 data (if NULL, data is returned instead)

overwrite
Should existing files be overwritten? Only relevant if outdir is not NULL

verbose
Print progress updates

Value

NULL if outdir is specified. Otherwise, it should return a list with f2 data arrays

Details

For each population pair, each of the \(i = 1, \ldots, n\) resutling values (\(n\) is around 700 in practice) is the mean \(f2\) estimate across all SNPs except the ones in block \(i\).

\(- p1 (1 - p1)/(2 n1 - 1) - p2 (1 - p2)/(2 n2 - 1)\) is a correction term which makes the estimates unbiased at low sample sizes.

Examples

if (FALSE) {
afdat = plink_to_aftable('/my/geno/prefix')
afs_to_f2_blocks(afdat, outdir = '/my/f2/data/')
}