Compute and store blocked f2 statistics

extract_f2_large does the same as extract_f2, but it requires less memory.

extract_f2_large(
  pref,
  outdir,
  inds = NULL,
  pops = NULL,
  dist = 0.05,
  cols_per_chunk = 10,
  maxmiss = 1,
  minmaf = 0,
  maxmaf = 0.5,
  transitions = TRUE,
  transversions = TRUE,
  keepsnps = NULL,
  snpblocks = NULL,
  overwrite = FALSE,
  format = NULL,
  poly_only = TRUE,
  adjust_pseudohaploid = TRUE,
  verbose = TRUE
)

Arguments

pref
Prefix of packedancestrymap or PLINK files. packedancestrymap has to end in .geno, .snp, .ind, PLINK has to end in .bed, .bim, .fam

outdir
Directory where data will be stored

inds
Individuals for which data should be extracted

pops
Populations for which data should be extracted. If both pops and inds are provided, they should have the same length and will be matched by position. If only pops is provided, all individuals from the .ind or .fam file in those populations will be extracted. If only inds is provided, each indivdual will be assigned to its own population of the same name. If neither pops nor inds is provided, all individuals and populations in the .ind or .fam file will be extracted.

dist
Genetic distance in Morgan. Default is 0.05 (50 cM).

cols_per_chunk
Number of populations per chunk. Lowering this number will lower the memory requirements when running

maxmiss
Discard SNPs which are missing in a fraction of populations higher than maxmiss

minmaf
Discard SNPs with minor allele frequency less than minmaf

maxmaf
Discard SNPs with minor allele frequency greater than than maxmaf

transitions
Set this to FALSE to exclude transition SNPs

transversions
Set this to FALSE to exclude transversion SNPs

keepsnps
SNP IDs of SNPs to keep. Overrides other SNP filtering options

snpblocks
Optional data frame with pre-assigned SNP blocks. Should have columns 'SNP' and 'block'. Can also have a column 'weight'. If present, the average 'weight' in each SNP block will be used for jackknife weights, instead of the number of SNPs in that block.

overwrite
Overwrite existing files in outdir

format
Supply this if the prefix can refer to genotype data in different formats and you want to choose which one to read. Should be either plink, ancestrymap, or packedancestrymap

poly_only
Exclude sites with identical allele frequencies in all populations

adjust_pseudohaploid
Genotypes of pseudohaploid samples are usually coded as 0 or 2, even though only one allele is observed. adjust_pseudohaploid ensures that the observed allele count increases only by 1 for each pseudohaploid sample. If TRUE (default), samples that don't have any genotypes coded as 1 among the first 1000 SNPs are automatically identified as pseudohaploid. This leads to slightly more accurate estimates of f-statistics. Setting this parameter to FALSE is equivalent to the ADMIXTOOLS inbreed: NO option.

verbose
Print progress updates

Value

SNP metadata (invisibly)

Details

extract_f2_large requires less memory because it writes allele frequency data to disk, and doesn't store the allele frequency matrix for all populations and SNPs in memory. If you still run out of memory, reduce cols_per_chunk. This function is a wrapper around extract_afs and write_split_f2_block, and is slower than extract_f2. It may be faster to call extract_afs and write_split_f2_block directly, parallelizing over the different calls to write_split_f2_block.

See also

Examples

if (FALSE) {
pref = 'my/genofiles/prefix'
f2dir = 'my/f2dir/'
extract_f2_large(pref, f2dir, pops = c('popA', 'popB', 'popC'))
}