Find well fitting admixture graphs

This function generates and evaluates admixture graphs in numgen iterations across numrep independent repeats to find well fitting admixturegraphs. It uses the function future_map from the furrr package to parallelize across the independent repeats. The function future::plan can be called to specify the details of the parallelization. This can be used to parallelize across cores or across nodes on a compute cluster. Setting numadmix to 0 will search for well fitting trees, which is much faster than searching for admixture graphs with many admixture nodes.

find_graphs(f2_data, pops = NULL, outpop = NULL, numrep = 10,
  numgraphs = 50, numgen = 5, numsel = 5, numadmix = 0,
  keep = c("all", "best", "last"), f2_denom = 1, cpp = TRUE,
  initgraph = NULL, mutfuns = c("spr_leaves", "spr_all", "swap_leaves",
  "move_admixedge_once", "flipadmix_random"), store_intermediate = NULL,
  parallel = TRUE, stop_at = NULL, debug = FALSE,
  fudge_cov = 1e-05, verbose = TRUE, ...)

Arguments

f2_data
A 3d array of blocked f2 statistics, output of f2_from_precomp. alternatively, a directory with precomputed data. see extract_f2 and extract_indpairs.

pops
Populations for which to fit admixture graphs

outpop
Outgroup population

numrep
Number of independent repetitions (each repetition can be run in parallel)

numgraphs
Number of graphs in each generation

numgen
Number of generations

numsel
Number of graphs which are selected in each generation. Should be less than numgraphs.

numadmix
Number of admixture events within each graph

keep
By default (all), all evaluated graphs will be returned. best will only return the best fitting graph from each repeat and each generation. last will return all graphs from the last generation.

f2_denom
Scales f2-statistics. A value of around 0.278 converts f2 to Fst.

cpp
Use C++ functions. Setting this to FALSE will be slower but can help with debugging.

initgraph
Optional graph to start with. If NULL, optimization will start with random graphs.

mutfuns
The names of funcations used to modify graphs.

spr_leaves Subtree prune and regraft leaves. Cuts a leaf node and attaches it to a random other edge in the graph.

spr_all Subtree prune and regraft. Cuts any edge and attaches the new orphan node to a random other edge in the graph, keeping the number of admixture nodes constant.

swap_leaves Swaps two leaf nodes.

move_admixedge_once Moves an admixture edge to a nearby location.

flipadmix_random Flips the direction of an admixture edge (if possible).

store_intermediate
Path and prefix of files for intermediate results to .rds. Can be useful if find_graphs doesn't finish sucessfully.

parallel
Parallelize over repeats (if numrep > 1) or graphs (if numrep == 1) by replacing purrr::map with furrr::future_map. Will only be effective if future::plan() has been set.

stop_at
Stop execution after finishing the generation running at stop_at seconds. Currently not working.

debug
If TRUE each repeat is run sequentially in a loop and not via furrr::map). Errors will interrupt execution. This is the default if numrep = 1

fudge_cov
Regularization term added to the covariance matrix of estimated f3 statistics (after scaling by the matrix trace).

verbose
Print progress updates

...
Additional arguments passed to qpgraph

Value

a nested data frame with one model per line

Examples

if (FALSE) {
find_graphs(example_f2_blocks, numrep = 200, numgraphs = 100,
            numgen = 20, numsel = 5, numadmix = 3)
}