Estimate f3 statistics

Computes f3 statistics of the form \(f3(A; B, C)\). When using the same SNPs for all populations, this is equivalent to \((f2(A, B) + f2(A, C) - f2(B, C)) / 2\) and to \(f4(A, B; A, C)\)

qp3pop(
  data,
  pop1 = NULL,
  pop2 = NULL,
  pop3 = NULL,
  boot = FALSE,
  sure = FALSE,
  unique_only = TRUE,
  blgsize = NULL,
  block_lengths = NULL,
  verbose = FALSE,
  ...
)

Arguments

data

Input data in one of three forms:

A 3d array of blocked f2 statistics, output of f2_from_precomp or f2_from_geno (fastest option)

A directory which contains pre-computed f2-statistics

The prefix of genotype files (slowest option)

pop1

One of the following four:

NULL: all possible population combinations will be returned

A vector of population labels. All combinations with the other pop arguments will be returned

A matrix with population combinations to be tested, with one population per column and one combination per row. Other pop arguments will be ignored.

the location of a file (poplistname or popfilename) which specifies the populations or population combinations to be tested. Other pop arguments will be ignored.

pop2

A vector of population labels

pop3

A vector of population labels

boot

If FALSE (the default), block-jackknife resampling will be used to compute standard errors. Otherwise, block-bootstrap resampling will be used to compute standard errors. If boot is an integer, that number will specify the number of bootstrap resamplings. If boot = TRUE, the number of bootstrap resamplings will be equal to the number of SNP blocks.

sure

The number of population combinations can get very large. This is a safety option that stops you from accidently computing all combinations if that number is large.

unique_only

If TRUE (the default), redundant combinations will be excluded

verbose

Print progress updates

...

Additional arguments passed to f3blockdat_from_geno if data is a genotype prefix, or to get_f2 otherwise

Value

qp3pop returns a data frame with f3 statistics

Details

There are several arguments that can be passed via ... which affect the estimated f3-statistics. The default options are the same as in the original qp3pop program, but some options are not effective when using precomputed f2-statistics. See examples for more information.

Alias

f3

References

Patterson, N. et al. (2012) Ancient admixture in human history Genetics

Peter, B. (2016) Admixture, Population Structure, and F-Statistics Genetics

Examples