Estimate admixture weights

Models target as a mixture of left populations, given a set of outgroup right populations. Can be used to estimate admixture proportions, and to estimate the number of independent admixture events (see wave).

qpadm(f2_data, target = NULL, left = NULL, right = NULL,
  wave = TRUE, f2_denom = 1, fudge = 1e-04, boot = FALSE,
  getcov = TRUE, constrained = FALSE, cpp = TRUE, verbose = TRUE)

Arguments

f2_data
a 3d array of blocked f2 statistics, output of f2_from_precomp. alternatively, a directory with precomputed data. see extract_f2 and extract_indpairs.

target
target population

left
source populations

right
outgroup populations

wave
If TRUE (the default), lower rank models and models with fewer source populations will also be tested.

f2_denom
scales f2-statistics. A value of around 0.278 converts F2 to Fst.

fudge
value added to diagonal matrix elements before inverting

boot
If FALSE (the default), each block will be left out at a time and the covariance matrix of f4 statistics, as well as the weight standard errors, will be computed using block-jackknife. Otherwise bootstrap resampling is performed n times, where n is either equal to boot if it is an integer, or equal to the number of blocks if boot is TRUE. The the covariance matrix of f4 statistics, as well as the weight standard errors, will be computed using bootstrapping.

getcov
should standard errors be returned? Setting this to FALSE makes this function much faster.

constrained
if FALSE (default), admixture weights can be negative. if TRUE, they will all be non-negative, as in lazadm

cpp
should optimization be done using C++ or R function? cpp = TRUE is much faster.

verbose
print progress updates

Value

a list with output describing the model fit:

weights a data frame with estimated admixture proportions where each row is left population. estimated edge length, and for admixture edges, it is the estimated admixture weight.

rankdrop a data frame describing model fits with different ranks, including p-values for the overall fit and for nested models (comparing two models with rank difference of one). returned only if wave = TRUE

popdrop a data frame describing model fits with different populations. returned only if wave = TRUE

Alias

qpwave

References

Patterson, N. et al. (2012) Ancient admixture in human history. Genetics

Haak, W. et al. (2015) Massive migration from the steppe was a source for Indo-European languages in Europe. Nature (SI 10)

See also

Examples

target = 'Denisova.DG'
left = c('Altai_Neanderthal.DG', 'Vindija.DG')
right = c('Chimp.REF', 'Mbuti.DG', 'Russia_Ust_Ishim.DG', 'Switzerland_Bichon.SG')
qpadm(example_f2_blocks, target, left, right)

#> ℹ Computing f4 stats...
#> ℹ Computing admixture weights...
#> ℹ Computing standard errors...
#> ℹ Computing number of admixture waves...
#> 

#> $weights
#> # A tibble: 2 x 5
#>   target      left                 weight    se     z
#>   <chr>       <chr>                 <dbl> <dbl> <dbl>
#> 1 Denisova.DG Altai_Neanderthal.DG   50.2  24.0  2.10
#> 2 Denisova.DG Vindija.DG            -49.2  24.0 -2.06
#> 
#> $f4
#> # A tibble: 36 x 9
#>    pop1    pop2       pop3    pop4            est      se     z         p weight
#>    <chr>   <chr>      <chr>   <chr>         <dbl>   <dbl> <dbl>     <dbl>  <dbl>
#>  1 Deniso… Altai_Nea… Chimp.… Mbuti.DG    0.0101  2.82e-4  35.6 4.79e-278   50.2
#>  2 Deniso… fit        Chimp.… Mbuti.DG   -0.00616 3.91e-4 -15.7 9.78e- 56   NA  
#>  3 Deniso… Vindija.DG Chimp.… Mbuti.DG    0.0102  2.90e-4  35.1 1.28e-269  -49.2
#>  4 Deniso… Altai_Nea… Chimp.… Russia_Us…  0.0118  3.48e-4  34.0 4.75e-253   50.2
#>  5 Deniso… fit        Chimp.… Russia_Us…  0.0144  8.93e-4  16.1 1.64e- 58   NA  
#>  6 Deniso… Vindija.DG Chimp.… Russia_Us…  0.0122  3.53e-4  34.5 4.40e-260  -49.2
#>  7 Deniso… Altai_Nea… Chimp.… Switzerla…  0.0116  3.65e-4  31.9 1.88e-223   50.2
#>  8 Deniso… fit        Chimp.… Switzerla…  0.0846  4.96e-4 171.  0.          NA  
#>  9 Deniso… Vindija.DG Chimp.… Switzerla…  0.0120  3.76e-4  31.9 1.14e-222  -49.2
#> 10 Deniso… Altai_Nea… Mbuti.… Chimp.REF  -0.0101  2.82e-4 -35.6 4.79e-278   50.2
#> # … with 26 more rows
#> 
#> $rankdrop
#> # A tibble: 1 x 7
#>   f4rank   dof chisq      p dofdiff chisqdiff p_nested
#> *  <int> <int> <dbl>  <dbl>   <int>     <dbl>    <dbl>
#> 1      1     2  7.14 0.0281      NA        NA       NA
#> 
#> $popdrop
#> # A tibble: 1 x 13
#>   pat      wt   dof chisq      p f4rank Altai_Neanderth… Vindija.DG feasible
#>   <chr> <dbl> <int> <dbl>  <dbl>  <int>            <dbl>      <dbl> <lgl>   
#> 1 00        0     2  7.14 0.0281      1             50.2      -49.2 FALSE   
#> # … with 4 more variables: best <lgl>, dofdiff <int>, chisqdiff <dbl>,
#> #   p_nested <dbl>
#>