Estimate admixture weights

Models target as a mixture of left populations, given a set of outgroup right populations. Can be used to estimate admixture proportions, and to estimate the number of independent admixture events.

qpadm(
  f2_data,
  left,
  right,
  target,
  fudge = 1e-04,
  boot = FALSE,
  getcov = TRUE,
  constrained = FALSE,
  cpp = TRUE,
  verbose = TRUE
)

Arguments

f2_data
A 3d array of blocked f2 statistics, output of f2_from_precomp. Alternatively, a directory with f2 statistics.

left
Left populations (sources)

right
Right populations (outgroups)

target
Target population

fudge
value added to diagonal matrix elements before inverting

boot
If FALSE (the default), each block will be left out at a time and the covariance matrix of f4 statistics, as well as the weight standard errors, will be computed using block-jackknife. Otherwise bootstrap resampling is performed n times, where n is either equal to boot if it is an integer, or equal to the number of blocks if boot is TRUE. The the covariance matrix of f4 statistics, as well as the weight standard errors, will be computed using bootstrapping.

getcov
Compute weights covariance. If not needed, turning this off will speed things up.

constrained
Constrain admixture weights to be non-negative

cpp
Use C++ functions. Setting this to FALSE will be slower but can help with debugging.

verbose
Print progress updates

Value

A list with output describing the model fit:

weights A data frame with estimated admixture proportions where each row is a left population.

f4 A data frame with estimated and fitted f4-statistics

rankdrop A data frame describing model fits with different ranks, including p-values for the overall fit and for nested models (comparing two models with rank difference of one).

popdrop A data frame describing model fits with different populations.

References

Haak, W. et al. (2015) Massive migration from the steppe was a source for Indo-European languages in Europe. Nature (SI 10)

See also

Examples

left = c('Altai_Neanderthal.DG', 'Vindija.DG')
right = c('Chimp.REF', 'Mbuti.DG', 'Russia_Ust_Ishim.DG', 'Switzerland_Bichon.SG')
target = 'Denisova.DG'
qpadm(example_f2_blocks, left, right, target)

#> ℹ Computing f4 stats...
#> ℹ Computing admixture weights...
#> ℹ Computing standard errors...
#> ℹ Computing number of admixture waves...
#> 

#> $weights
#> # A tibble: 2 x 5
#>   target      left                 weight    se     z
#>   <chr>       <chr>                 <dbl> <dbl> <dbl>
#> 1 Denisova.DG Altai_Neanderthal.DG   49.6  23.3  2.13
#> 2 Denisova.DG Vindija.DG            -48.6  23.3 -2.08
#> 
#> $f4
#> # A tibble: 36 x 9
#>    pop1    pop2       pop3   pop4           est      se       z         p weight
#>    <chr>   <chr>      <chr>  <chr>        <dbl>   <dbl>   <dbl>     <dbl>  <dbl>
#>  1 Deniso… Altai_Nea… Chimp… Mbuti.DG   0.0129  3.64e-4  35.6   2.22e-277   49.6
#>  2 Deniso… fit        Chimp… Mbuti.DG   0.00693 6.60e-3   1.05  2.94e-  1   NA  
#>  3 Deniso… Vindija.DG Chimp… Mbuti.DG   0.0131  3.73e-4  35.0   7.55e-269  -48.6
#>  4 Deniso… Altai_Nea… Chimp… Russia_U…  0.0152  4.46e-4  34.0   4.67e-254   49.6
#>  5 Deniso… fit        Chimp… Russia_U… -0.00642 8.03e-3  -0.800 4.23e-  1   NA  
#>  6 Deniso… Vindija.DG Chimp… Russia_U…  0.0156  4.53e-4  34.5   2.14e-261  -48.6
#>  7 Deniso… Altai_Nea… Chimp… Switzerl…  0.0150  4.64e-4  32.3   6.06e-229   49.6
#>  8 Deniso… fit        Chimp… Switzerl… -0.00552 8.43e-3  -0.654 5.13e-  1   NA  
#>  9 Deniso… Vindija.DG Chimp… Switzerl…  0.0154  4.78e-4  32.2   5.81e-228  -48.6
#> 10 Deniso… Altai_Nea… Mbuti… Chimp.REF -0.0129  3.64e-4 -35.6   2.22e-277   49.6
#> # … with 26 more rows
#> 
#> $rankdrop
#> # A tibble: 2 x 7
#>   f4rank   dof   chisq      p dofdiff chisqdiff p_nested
#> *  <int> <int>   <dbl>  <dbl>   <int>     <dbl>    <dbl>
#> 1      1     2    7.15 0.0280       4     1572.        0
#> 2      0     6 1580.   0           NA       NA        NA
#> 
#> $popdrop
#> # A tibble: 3 x 13
#>   pat      wt   dof  chisq      p f4rank Altai_Neanderth… Vindija.DG feasible
#>   <chr> <dbl> <dbl>  <dbl>  <dbl>  <dbl>            <dbl>      <dbl> <lgl>   
#> 1 00        0     2 7.15e0 0.0280      1             49.6      -48.6 FALSE   
#> 2 01        1     3 1.14e4 0           0              1         NA   TRUE    
#> 3 10        1     3 1.14e4 0           0             NA          1   TRUE    
#> # … with 4 more variables: best <lgl>, dofdiff <dbl>, chisqdiff <dbl>,
#> #   p_nested <dbl>
#>