Estimate f4 statistics

Computes f4 statistics from f2 blocks of the form \(f4(A, B; C, D)\). Equivalent to \((f2(A, D) + f2(B, C) - f2(A, C) - f2(B, D)) / 2\)

qpdstat(f2_data, pop1 = NULL, pop2 = NULL, pop3 = NULL,
  pop4 = NULL, f2_denom = 1, boot = FALSE, sure = FALSE,
  unique_only = TRUE, verbose = FALSE)

Arguments

f2_data
a 3d array of block-jackknife leave-one-block-out estimates of f2 statistics, output of afs_to_f2_blocks. alternatively, a directory with f2 statistics. see extract_indpairs.

pop1
one of the following four:

NULL: all possible pairs of the populations in f2_blocks will be returned

a vector of population labels

a data frame with population combinations to be tested, with one population per column and one combination per row. Other pop arguments will be ignored.

the location of a file (poplistname or popfilename) which specifies the populations or population combinations to be tested. Other pop arguments will be ignored.

pop2
a vector of population labels

pop3
a vector of population labels

pop4
a vector of population labels

f2_denom
scales f2-statistics. 1 correspondes to f4mode: YES. 1/4.75 is similar to f4mode: NO.

boot
If FALSE (the default), each block will be left out at a time and the covariance matrix of the f statistics will be computed using block-jackknife. Otherwise bootstrap resampling is performed n times, where n is either equal to boot if it is an integer, or equal to the number of blocks if boot is TRUE. The the covariance matrix of the f statistics will be computed using bootstrapping.

sure
The number of population combinations can get very large. This is a safety option that stops you from accidently computing all combinations if that number is large.

unique_only
If TRUE (the default), redundant combinations will be returned as well.

verbose
print progress updates

Value

a data frame with f4 statistics

Alias

f4

References

Patterson, N. et al. (2012) Ancient admixture in human history. Genetics

Examples

pop1 = 'Denisova.DG'
pop2 = c('Altai_Neanderthal.DG', 'Vindija.DG')
pop3 = c('Chimp.REF', 'Mbuti.DG', 'Russia_Ust_Ishim.DG')
pop4 = 'Switzerland_Bichon.SG'
qpdstat(example_f2_blocks, pop1, pop2, pop3, pop4)

#> # A tibble: 6 x 8
#>   pop1     pop2        pop3       pop4             est      se       z         p
#>   <chr>    <chr>       <chr>      <chr>          <dbl>   <dbl>   <dbl>     <dbl>
#> 1 Denisov… Altai_Nean… Chimp.REF  Switzerlan…  1.16e-2 3.65e-4  31.9   1.88e-223
#> 2 Denisov… Altai_Nean… Mbuti.DG   Switzerlan…  1.58e-3 2.76e-4   5.72  1.04e-  8
#> 3 Denisov… Altai_Nean… Russia_Us… Switzerlan… -1.68e-4 2.93e-4  -0.575 5.65e-  1
#> 4 Denisov… Vindija.DG  Chimp.REF  Switzerlan…  1.20e-2 3.76e-4  31.9   1.14e-222
#> 5 Denisov… Vindija.DG  Mbuti.DG   Switzerlan…  1.81e-3 2.83e-4   6.39  1.67e- 10
#> 6 Denisov… Vindija.DG  Russia_Us… Switzerlan… -1.86e-4 3.04e-4  -0.614 5.39e-  1

if (FALSE) {
qpdstat(f2_dir, pop1, pop2, pop3, pop4)
}