Estimate f4 statistics

Computes f4 statistics from f2 blocks of the form \(f4(A, B; C, D)\). Equivalent to \((f2(A, D) + f2(B, C) - f2(A, C) - f2(B, D)) / 2\)

qpdstat(
  data,
  pop1 = NULL,
  pop2 = NULL,
  pop3 = NULL,
  pop4 = NULL,
  boot = FALSE,
  sure = FALSE,
  unique_only = TRUE,
  comb = TRUE,
  blgsize = NULL,
  block_lengths = NULL,
  f4mode = TRUE,
  afprod = TRUE,
  cpp = TRUE,
  verbose = is.character(data)
)

Arguments

data
Input data in one of three forms:

A 3d array of blocked f2 statistics, output of f2_from_precomp or extract_f2 (fastest option)

A directory which contains pre-computed f2-statistics

The prefix of genotype files (slowest option)

pop1
One of the following four:

NULL: all possible population combinations will be returned

A vector of population labels. All combinations with the other pop arguments will be returned

A matrix with population combinations to be tested, with one population per column and one combination per row. Other pop arguments will be ignored.

the location of a file (poplistname or popfilename) which specifies the populations or population combinations to be tested. Other pop arguments will be ignored.

pop2
A vector of population labels

pop3
A vector of population labels

pop4
A vector of population labels

boot
If FALSE (the default), each block will be left out at a time and the covariance matrix of the f statistics will be computed using block-jackknife. Otherwise bootstrap resampling is performed n times, where n is either equal to boot if it is an integer, or equal to the number of blocks if boot is TRUE. The the covariance matrix of the f statistics will be computed using bootstrapping.

sure
The number of population combinations can get very large. This is a safety option that stops you from accidently computing all combinations if that number is large.

unique_only
If TRUE (the default), redundant combinations will be excluded

comb
Generate all combinations of pop1, pop2, pop3, pop4. If FALSE, pop1, pop2, pop3, pop4 should all be vectors of the same length.

blgsize
SNP block size in Morgan. Default is 0.05 (50 cM). Only used when data is the prefix of genotype files

block_lengths
Vector with lengths of each jackknife block. sum(block_lengths) has to match the number of SNPs. only used when data is the prefix of genotype files

f4mode
If TRUE: f4 is computed from allele frequencies a, b, c, and d as (a-b)*(c-d). if FALSE, D-statistics are computed instead, defined as (a-b)*(c-d) / ((a + b - 2*a*b) * (c + d - 2*c*d)), which is the same as (P(ABBA) - P(BABA)) / (P(ABBA) + P(BABA)). f4mode = FALSE is only available when data is the prefix of genotype files

afprod
Compute f4 from allele frequency products instead of f2. Only used if data is a directory with precomputed data.

cpp
Use C++ functions. Setting this to FALSE will be slower but can help with debugging.

verbose
Print progress updates

Value

qpdstat returns a data frame with f4 statistics

Alias

f4

References

Patterson, N. et al. (2012) Ancient admixture in human history Genetics

Peter, B. (2016) Admixture, Population Structure, and F-Statistics Genetics

Examples

pop1 = 'Denisova.DG'
pop2 = c('Altai_Neanderthal.DG', 'Vindija.DG')
pop3 = c('Chimp.REF', 'Mbuti.DG', 'Russia_Ust_Ishim.DG')
pop4 = 'Switzerland_Bichon.SG'
qpdstat(example_f2_blocks, pop1, pop2, pop3, pop4)

#> # A tibble: 6 x 8
#>   pop1     pop2        pop3        pop4             est      se      z         p
#>   <chr>    <chr>       <chr>       <chr>          <dbl>   <dbl>  <dbl>     <dbl>
#> 1 Denisov… Altai_Nean… Chimp.REF   Switzerlan…  1.50e-2 4.64e-4 32.3   6.06e-229
#> 2 Denisov… Altai_Nean… Mbuti.DG    Switzerlan…  2.03e-3 3.53e-4  5.75  8.85e-  9
#> 3 Denisov… Altai_Nean… Russia_Ust… Switzerlan… -2.17e-4 3.73e-4 -0.580 5.62e-  1
#> 4 Denisov… Vindija.DG  Chimp.REF   Switzerlan…  1.54e-2 4.78e-4 32.2   5.81e-228
#> 5 Denisov… Vindija.DG  Mbuti.DG    Switzerlan…  2.33e-3 3.63e-4  6.42  1.40e- 10
#> 6 Denisov… Vindija.DG  Russia_Ust… Switzerlan… -2.40e-4 3.87e-4 -0.620 5.35e-  1

if (FALSE) {
qpdstat(f2_dir, pop1, pop2, pop3, pop4)
}