Estimate admixture waves

qpwave compares two sets of populations (left and right) to each other. It estimates a lower bound on the number of admixtue waves that went from left into right, by comparing a matrix of f4-statistics to low-rank approximations. For a rank of 0 this is equivalent to testing whether left and right form clades relative to each other.

qpwave(
  f2_data,
  left,
  right,
  fudge = 1e-04,
  boot = FALSE,
  constrained = FALSE,
  cpp = TRUE,
  verbose = TRUE
)

Arguments

left
Left populations (sources)

right
Right populations (outgroups)

fudge
value added to diagonal matrix elements before inverting

boot
If FALSE (the default), each block will be left out at a time and the covariance matrix of f4 statistics, as well as the weight standard errors, will be computed using block-jackknife. Otherwise bootstrap resampling is performed n times, where n is either equal to boot if it is an integer, or equal to the number of blocks if boot is TRUE. The the covariance matrix of f4 statistics, as well as the weight standard errors, will be computed using bootstrapping.

constrained
Constrain admixture weights to be non-negative

cpp
Use C++ functions. Setting this to FALSE will be slower but can help with debugging.

verbose
Print progress updates

Value

A list with output describing the model fit:

f4 A data frame with estimated f4-statistics

rankdrop A data frame describing model fits with different ranks, including p-values for the overall fit and for nested models (comparing two models with rank difference of one).

References

Patterson, N. et al. (2012) Ancient admixture in human history. Genetics

Haak, W. et al. (2015) Massive migration from the steppe was a source for Indo-European languages in Europe. Nature (SI 10)

See also

Examples

left = c('Altai_Neanderthal.DG', 'Vindija.DG')
right = c('Chimp.REF', 'Mbuti.DG', 'Russia_Ust_Ishim.DG', 'Switzerland_Bichon.SG')
qpwave(example_f2_blocks, left, right)

#> ℹ Computing f4 stats...
#> ℹ Computing number of admixture waves...
#> 

#> $f4
#> # A tibble: 3 x 8
#>   pop1           pop2     pop3     pop4                est      se     z       p
#>   <chr>          <chr>    <chr>    <chr>             <dbl>   <dbl> <dbl>   <dbl>
#> 1 Altai_Neander… Vindija… Chimp.R… Mbuti.DG        1.24e-4 1.35e-4 0.920 0.358  
#> 2 Altai_Neander… Vindija… Chimp.R… Russia_Ust_Ish… 4.45e-4 1.64e-4 2.72  0.00653
#> 3 Altai_Neander… Vindija… Chimp.R… Switzerland_Bi… 4.22e-4 1.72e-4 2.45  0.0144 
#> 
#> $rankdrop
#> # A tibble: 1 x 7
#>   f4rank   dof chisq       p dofdiff chisqdiff p_nested
#>    <int> <int> <dbl>   <dbl>   <int>     <dbl>    <dbl>
#> 1      0     3  11.9 0.00768      NA        NA       NA
#>